pyproject.toml
[tool.poetry]
name = "kernel-analytic"
version = "0.0"
description = "Description of your tool"
authors = ["Marco Simoni <marco.simoni@iit.cnr.it>"]
repository = "https://github.com/sifis-home/Kernel-Analytic"
license = "MIT"
classifiers = [
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT",
    "Natural Language :: English",
    "Programming Language :: Python :: 3 :: Only",
]
include = ["README.md", "LICENSE.md", "kernel-analytic/data/*"]
packages = [
    { include = "kernel-analytic" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
docker = "^6.1.3"
websocket-client = "1.4.2"
pytest = "7.4.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
pre-commit = "^3.0.4"
black = "^23.1.0"
isort = "^5.12.0"
ruff = "^0.0.63"
coverage = {extras = ["toml"], version = "^7.1.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.ruff]
line-length = 79
select = [
    "F401",
    "F403",
]

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 60
